+++ src/main/java/org/grnet/pidmr/dto/ResolutionModeRequestV2.java
@@ -7,7 +7,7 @@ import org.eclipse.microprofile.openapi.annotations.enums.SchemaType;
 import org.eclipse.microprofile.openapi.annotations.media.Schema;
 import org.grnet.pidmr.entity.database.Endpoint;
 
-import java.util.List;
+import java.util.Set;
 
 @Schema(name="ResolutionModeRequestV2", description="Request to create a resolution mode.")
 public class ResolutionModeRequestV2 {
@@ -27,5 +27,5 @@ public class ResolutionModeRequestV2 {
             description = "The set of resolution mode endpoints."
     )
     @JsonInclude(JsonInclude.Include.NON_NULL)
-    public List<@Valid Endpoint> endpoints;
+    public Set<@Valid Endpoint> endpoints;
 }