+++ src/main/java/org/grnet/pidmr/enums/ProviderStatus.java
@@ -3,5 +3,6 @@ package org.grnet.pidmr.enums;
 public enum ProviderStatus {
 
     PENDING,
-    APPROVED
+    APPROVED,
+    REJECTED
 }