+++ src/main/java/org/grnet/pidmr/service/ProviderService.java
@@ -3,8 +3,8 @@ package org.grnet.pidmr.service;
 import com.fasterxml.jackson.core.json.JsonReadFeature;
 import com.fasterxml.jackson.databind.json.JsonMapper;
 import jakarta.enterprise.context.ApplicationScoped;
+import jakarta.inject.Inject;
 import jakarta.inject.Named;
-import jakarta.transaction.Transactional;
 import jakarta.ws.rs.NotAcceptableException;
 import jakarta.ws.rs.NotSupportedException;
 import jakarta.ws.rs.core.UriInfo;
@@ -43,6 +43,8 @@ public class ProviderService implements ProviderServiceI {
     @ConfigProperty(name = "list.actions.file")
     String actionsPath;
 
+    @Inject
+    Utility utility;
 
     /**
      * This method is responsible for paginating the available Providers.
@@ -57,7 +59,7 @@ public class ProviderService implements ProviderServiceI {
 
         var allProviders = getProviders();
 
-        var partition = Utility.partition(new ArrayList<>(allProviders), size);
+        var partition = utility.partition(new ArrayList<>(allProviders), size);
 
         var providers = partition.get(page) == null ? Collections.EMPTY_LIST : partition.get(page);
 
@@ -90,7 +92,7 @@ public class ProviderService implements ProviderServiceI {
                 .enable(JsonReadFeature.ALLOW_BACKSLASH_ESCAPING_ANY_CHARACTER)
                 .build();
 
-        return Utility.toSet(Provider.class, mapper, providersPath);
+        return utility.toSet(Provider.class, mapper, providersPath);
     }
 
     /**
@@ -104,7 +106,7 @@ public class ProviderService implements ProviderServiceI {
                 .builder()
                 .build();
 
-        return Utility.toSet(Action.class, mapper, actionsPath);
+        return utility.toSet(Action.class, mapper, actionsPath);
     }
 
     /**